+++ test-data/src/main/java/org/ehrbase/test_data/composition/CompositionTestDataCanonicalJson.java
@@ -60,7 +60,9 @@ public enum CompositionTestDataCanonicalJson {
     MINIMAL_WITH_OPTIONAL_ATTRIBUTE("dv multimedia with alternate_text", "minimal_with_optional_attribute.json"),
     MINIMAL_WITHOUT_OPTIONAL_ATTRIBUTE("dv multimedia without alternate_text", "minimal_without_optional_attribute.json"),
     GECCO_PERSONENDATEN("GECCO_Personendaten", "gecco_personendaten.json"),
-    GECCO_LABORBEFUND("GECCO_Laborbefund", "gecco_laborbefund.json");
+    GECCO_LABORBEFUND("GECCO_Laborbefund", "gecco_laborbefund.json"),
+    PARTICIPATION_NO_CONTENT("to test various participation CR #710", "participation_no_content.json"),
+    IPS("Internation Patient Summary", "ips_canonical.json");
 
 
     private final String filename;