+++ web-template/src/main/java/org/ehrbase/webtemplate/filter/Filter.java
@@ -19,9 +19,20 @@
 
 package org.ehrbase.webtemplate.filter;
 
+import static org.ehrbase.webtemplate.parser.OPTParser.DV_CODED_TEXT;
+
 import com.nedap.archie.rm.datastructures.Event;
 import com.nedap.archie.rminfo.ArchieRMInfoLookup;
 import com.nedap.archie.rminfo.RMTypeInfo;
+import java.util.ArrayDeque;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Deque;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.stream.Collectors;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.commons.lang3.tuple.Pair;
 import org.ehrbase.util.reflection.ReflectionHelper;
@@ -33,11 +44,7 @@ import org.ehrbase.webtemplate.model.WebTemplateNode;
 import org.ehrbase.webtemplate.parser.InputHandler;
 import org.ehrbase.webtemplate.parser.OPTParser;
 import org.ehrbase.webtemplate.parser.config.RmIntrospectConfig;
-
-import java.util.*;
-import java.util.stream.Collectors;
-
-import static org.ehrbase.webtemplate.parser.OPTParser.DV_CODED_TEXT;
+import org.ehrbase.webtemplate.util.WebTemplateUtils;
 
 public class Filter implements WebTemplateFilter {
 
@@ -108,27 +115,17 @@ public class Filter implements WebTemplateFilter {
     }
 
     if (node.getRmType().equals("ELEMENT")) {
-      List<WebTemplateNode> trueChildren =
-          node.getChildren().stream()
-              .filter(n -> !"name".equals(n.getName()))
-              .filter(
-                  n ->
-                      !List.of("null_flavour", "feeder_audit").contains(n.getName())
-                          || !n.isNullable())
-              .collect(Collectors.toList());
-      if (trueChildren.stream()
-              .map(WebTemplateNode::getRmType)
-              .collect(Collectors.toList())
-              .containsAll(List.of("DV_TEXT", DV_CODED_TEXT))
-          && node.getChoicesInChildren().size() > 0
-          && trueChildren.size() == 2) {
+      if (WebTemplateUtils.isChoiceDvCodedTextAndDvText(node)) {
         WebTemplateNode merged = mergeDVText(node);
         merged.setId(node.getId());
         node.getChildren().clear();
         node.getChildren().add(merged);
-      } else if (trueChildren.size() == 1) {
-        // Element will be skipped and the value node inherits the id
-        trueChildren.get(0).setId(node.getId());
+      } else {
+        List<WebTemplateNode> trueChildren = WebTemplateUtils.getTrueChildrenElement(node);
+        if (trueChildren.size() == 1) {
+          // Element will be skipped and the value node inherits the id
+          trueChildren.get(0).setId(node.getId());
+        }
       }
     }
   }